Add ability to configure settings that output as CSS variables
Added a migration that creates a new `ThemeSettings` content type with
a stereotype of custom settings. Currently it only contains a field that
represents the primary colour. These settings are then utilised by a
result filter to convert the theme settings in to CSS variables that are
included in the page render.